+++ options.js
@@ -1,1 +1,1 @@
-export { options as default } from './lib/cli.js'
+export { default, resolveOptions } from './lib/options.js'