+++ proxy.js
@@ -1,4 +1,4 @@
-import { options as cliOptions } from './lib/cli.js'
+import getOptions from './lib/options.js'
 import proxy from './lib/proxy.js'
 import createLogger from './lib/logger.js'
 
@@ -6,7 +6,7 @@ let logPromise
 
 function getLog() {
   if (!logPromise) {
-    logPromise = cliOptions().then((opts) => createLogger(opts.logLevels))
+    logPromise = getOptions().then((opts) => createLogger(opts.logLevels))
   }
   return logPromise
 }